requirements.txt
Django==4.2
channels==4.0.0
uvicorn==0.22.0
sqlparse==0.4.3
streamlit==1.26.0
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.4.0
